refactor: moved resource read logic into a new module in root. (#1622)
